main.cpp書き込み前実行確認したら警告出たので修正。 ・main.cppでTextureManagerクラスのGetImage関数をGetImagePtr関数に変更した部分の修正。 ・Sprite.cppのDrawCall部分微修正。 ・VertexBuffrt.hでGetVerticesNum()がuintのままだと2回型変換することになるので1回で済むように調整。 後なんかキーボード入力の時くそ重かった。

+++ Sprite.cpp
@@ -74,7 +74,7 @@ void Sprite::Draw(void)
     iDXPtr->GetCommandList()->SetGraphicsRootDescriptorTable(0, imagePtr_->srvGpuHandle_);
 
     // 描画コマンド
-    iDXPtr->GetCommandList()->DrawInstanced(vertexBuffer_->GetVerticesNum(), 1, 0, 0); // 全ての頂点を使って描画
+    iDXPtr->GetCommandList()->DrawInstanced((unsigned int)vertexBuffer_->GetVerticesNum(), 1, 0, 0); // 全ての頂点を使って描画
 }
 
 void Sprite::TransferVertex(void)
